Check for zero code coverage explicitly
When a file comes back with no code coverage, the format will not be as expected
by the parseXMLCoverageData function. This teaches it how to handle that
case (by simplying returning an empty coverage object).

Also, when comparing the coverage a particular spec, the base coverage might
have no coverage at all. Treat this as zero coverage rather than undefined
coverage, which makes the comparison logic work better.