Added various karma plugins to dependencies
This package seems to need an old version of karma to execute, so we want our
plugins to match. If we let it come from the host environment, it'll likely be a
newer version that doesn't work with karma 0.13.21.

Note: This is solving particular problems for me, and probably isn't generally
useful. Use at your own risk.